+++ scripts/levels.js
@@ -18,10 +18,10 @@ export let level1 = [
 	"..w.k.........w...................................",
 	"..w...........w...................................",
 	"..wwwww.D.wwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwww",
-	"..w..................m.....w.....................w",
+	"..w..................M.....w.....................w",
 	"..w..............................................w",
 	"..w..........k.............d..............x......w",
-	"..w...m..........................................w",
+	"..w...M..........................................w",
 	"..w........................w.....................w",
 	"..wwwwwwwwwwwwwwwwwwwwwwwwww.....................w",
 	"...........................wwwwwwwwwwwwwwwwwwwwwww",
